backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
requests>=2.31,<3
beautifulsoup4>=4.12
ddgs>=6.0
python-multipart>=0.0.9
pypdf>=4.0
python-docx>=1.0
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
